+++ doc/CHANGELOG.md
@@ -7,6 +7,20 @@ permalink: changelog.html
 summary: The Changelog of The &tau; Project.
 toc: false
 ---
+### 9.20.0
+
+- On iOS : codec==Codec.pcm16 and codec==Codec.pcm16WAV  --  startRecorder()  --   Parameter `numChannels` is correctely handled (if >1 then the audio samples are interleaved)
+- On iOS : codec==Codec.pcm16 and codec==Codec.pcm16WAV  --  startPlayer()    --   Parameter `numChannels` is correctely handled (if >1 then the audio samples are interleaved)
+- On iOS : codec==Codec.pcm16 and codec==Codec.pcm16WAV  --  startRecorder()  --   Parameter `sampleRate`  is correctely handled
+- On iOS : codec==Codec.pcm16 and codec==Codec.pcm16WAV  --  startPlayer()    --   Parameter `sampleRate`  is correctely handled
+
+- On iOS : The peak level during recording pcm16 was unstable.
+- On iOS : codec==Codec.pcm16WAV  --  startRecorder()  --  The frames are correctely coded with int16 and not float32.
+- FlutterSoundHelper::pcmToWaveBuffer() : Add parameter Codec and implement codec==Codec.pcmFloat32
+
+- On iOS : codec==Codec.pcmFloat32 and codec==Codec.pcmFloat32WAV  --  startRecorder()    --  Float32 is implemented 
+- On iOS : codec==Codec.pcmFloat32 and codec==Codec.pcmFloat32WAV  --  startPlayer()    --   Float32 is implemented
+
 ### 9.19.1
 
 - flutter_sound_web depends on etau and tau_web